# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.GetProfileResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      session_token:
        type: string
      user:
        $ref: '#/definitions/auth.LoginUserResponse'
    type: object
  auth.LoginUserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
  auth.RegisterResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
...
  chat.GetChannelResponse:
    properties:
      created_by:
        type: integer
      id:
        type: integer
      name:
        type: string
      project_id:
        type: integer
      type:
        type: string
      workplace_id:
        type: integer
    type: object
...
  internal_handler_rest_v1_organizations.OrganizationWithRoleResponse:
    properties:
      id:
        type: integer
      joined_at:
        type: string
      logo_url:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/model.OrgRole'
    type: object
  model.OrgRole:
    enum:
    - admin
    - sub_admin
    - member
    type: string
    x-enum-varnames:
    - OrgRoleAdmin
    - OrgRoleSubAdmin
    - OrgRoleMember
  organizations.CreateOrganizationRequest:
    properties:
      description:
        type: string
      logo_url:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  organizations.CreateOrganizationResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      logo_url:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  projects.GetProjectResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      workplace_id:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for DPC Tools.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DPC Tools API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return session token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Login
      tags:
      - auth
  /auth/profile:
    get:
      description: Get the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.GetProfileResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.RegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Registration
      tags:
      - auth
...
  /channels/{id}:
    get:
      description: Retrieve details of a specific chat channel by its ID
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/chat.GetChannelResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Chat Channel
      tags:
      - chat
...
  /organizations:
    get:
      description: Retrieve all organizations the current user is a member of
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handler_rest_v1_organizations.OrganizationWithRoleResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List Organizations by User
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Create a new organization and assign the creator as ADMIN
      parameters:
      - description: Organization details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/organizations.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/organizations.CreateOrganizationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create Organization
      tags:
      - organizations
...
  /projects/{id}:
    get:
      description: Retrieve details of a specific project by its ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/projects.GetProjectResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Project
      tags:
      - projects
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"